requirements.txt
fastapi==0.100.0

fastapi-mail==1.4.1
fastapi-sessions==0.3.2
frozenlist==1.4.0
gunicorn==21.2.0

pydantic==2.0.2

PyJWT==2.8.0

python-multipart==0.0.6
SQLAlchemy==2.0.18
starlette==0.27.0

uvicorn==0.22.0
uvloop==0.17.0
websockets==11.0.3

zstandard==0.21.0
